+++ client/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.12.2
 
+- Update `avail-light-core` to 1.0.4
 - Add back origin attribute to metrics
 
 ## [1.12.1](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.12.1) - 2024-10-22